Rename the PFW.Model.Game namespace to PFW.Model.Match.
