+++ src/FieldWarning/Assets/Camera/SlidingCameraBehaviour.cs
@@ -16,7 +16,7 @@ using System;
 using UnityEngine;
 using UnityEngine.UI;
 
-using PFW.Model.Game;
+using PFW.Model.Match;
 
 namespace PFW
 {
