+++ src/FieldWarning/Assets/UI/Ingame/CaptureZone/CaptureZone.cs
@@ -11,7 +11,7 @@
  * the License for the specific language governing permissions and limitations under the License.
  */
 
-using PFW.Model.Game;
+using PFW.Model.Match;
 
 using System.Collections.Generic;
 using UnityEngine;
