+++ src/FieldWarning/Assets/UI/Ingame/TeamButton.cs
@@ -13,7 +13,7 @@
 
 using UnityEngine;
 
-using PFW.Model.Game;
+using PFW.Model.Match;
 
 public class TeamButton : MonoBehaviour
 {
