+++ src/FieldWarning/Assets/Units/MobilityData.cs
@@ -15,7 +15,7 @@ using UnityEngine;
 using System.Collections.Generic;
 
 using PFW.Model.Armory.JsonContents;
-using PFW.Model.Game;
+using PFW.Model.Match;
 using PFW.Units;
 using PFW.Units.Component.Movement;
